% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSignature.R
\name{predictSignature}
\alias{predictSignature}
\title{create svm model to predict yy from xx and report the properties}
\usage{
predictSignature(
  genesUse = genesUse,
  geneNames = geneNames,
  grpInd,
  log2out = log2out,
  cv = TRUE,
  rocPlot = TRUE
)
}
\arguments{
\item{genesUse}{vector of gene names comprising the signature}

\item{geneNames}{vector of all gene names correspondint to the rows of data}

\item{log2out}{log2 transformed data frame of protein abundances}

\item{cv}{if this is true, calculate cross-validation error rates}

\item{rocPlot}{if this is true, show an ROC plot}
}
\value{
result a dataframe with properties of the predictive model
}
\description{
create svm model to predict yy from xx and report the properties
}
