% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmProt.R
\name{svmProt}
\alias{svmProt}
\title{create svm model to predict yy from xx and report the properties}
\usage{
svmProt(xx, yy)
}
\arguments{
\item{xx}{data frame of data}

\item{yy}{actual reponse variable to be predicted}
}
\value{
result data structure with properties
}
\description{
create svm model to predict yy from xx and report the properties
}
