% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmProt2.R
\name{svmProt2}
\alias{svmProt2}
\title{create svm model for crossvalidation; for internal use}
\usage{
svmProt2(XTrain, YTrain, XTest, YTest)
}
\arguments{
\item{XTrain}{training data}

\item{YTrain}{response for training data}

\item{XTest}{test data}

\item{YTest}{response for test data}
}
\value{
result data structure with properties
}
\description{
create svm model for crossvalidation; for internal use
}
